Move StreamRecorder to testing package.
