+++ android-interop-testing/app/build.gradle
@@ -68,5 +68,6 @@ dependencies {
     compile 'io.grpc:grpc-protobuf-nano:0.9.0-SNAPSHOT'
     compile 'io.grpc:grpc-okhttp:0.9.0-SNAPSHOT'
     compile 'io.grpc:grpc-stub:0.9.0-SNAPSHOT'
+    compile 'io.grpc:grpc-testing:0.9.0-SNAPSHOT'
     compile 'javax.annotation:javax.annotation-api:1.2'
 }
