+++ testing/src/main/java/io/grpc/testing/StreamRecorder.java
@@ -29,10 +29,12 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.grpc.stub;
+package io.grpc.testing;
 
 import com.google.common.util.concurrent.ListenableFuture;
 import com.google.common.util.concurrent.SettableFuture;
+
+import io.grpc.stub.StreamObserver;
 
 import java.util.ArrayList;
 import java.util.Collections;
